ios update
The Podfile and lockfile have all changed since the dependnecies have
changed so much (because of the AndroidX update). I've had to comment
out the flutter_email_sender for now, since I just cannot seem to get
it to work, in fact, even the current build is a hit and miss.