ios: Enable flutter_email_sender
It needed the SWIFT_VERSION variable to be set, and for it to be set to
a version that exists on my system. I'm pasting the error over here so
maybe I google search will help a lost soul in the future -

`flutter_email_sender` does not specify a Swift version and none of the targets (`Runner`) integrating it have the `SWIFT_VERSION` attribute set. Please contact the author or set the `SWIFT_VERSION` attribute in at least one of the targets that integrate this pod.

https://github.com/sidlatau/flutter_email_sender/issues/8