Add a 'dev' and 'prod' build flavour
Now whenever we run 'flutter run' or 'flutter build' we need to pass the
flavour flag. It seems there is no way to set a default flavour.

This is required as I want to be able to not have to reconfigure my
local GitJournal app (which I use for note-taking) vs the one I'm
developing where I might change many different settings for the basis of
testing.

Specially since I don't use Github/GitLab to store my notes, so the
setup process is a bit more involved.