Surround the title by quotes before saving it to yaml
Otherwise people can put # or : or something else in the title and that
will result in invalid YAML.

Also escape any existing '. YAML is weird that the method of escaping a
' is by specifying it twice. It's sad that there is no dart YAML
serializer, and I'm surely missing some other edge cases. The YAML
specification is quite long.