Remove the BROWSE title
It's not required. Yes, stuff looks a little bit empty - but the title
over here really serves no purpose. Also, it's a bit confusing when
coming in from the hero animation for the List Tile's title to become
"BROWSE".

I do wish there was an easy way to inform the user that they can swipe
left/right in order to move between notes.