Avoid loading all the Notes in one go
This can sometimes lead to too many files being opened in parallel and
then failing because there aren't any file descriptors left. This patch
isn't perfect as given the directory structure we could still be loading
too many in one go, but it's a quick workaround for now.

The proper solution is to only load the Notes which need to be displayed
instead of loading all of them at startup.