Add a PoC for rendering equations with KaTex
It kind of works. The main problem is that the webview shows itself and
doesn't disappear after taking a screenshot. Also, I'm having to use an
old form of the community flutter webview plugin as the official
"community" plugin, doesn't support screenshots.

Overall, not bad. It'll probably just be easier to fork the webview
completely and fix the issues I'm encountering.

on iOS; Trying to take a screenshot resulted in a crash.