Setup: Move generating ssh key + deploy key code out of widget
It's strange, but this widget routinely gets constructed twice. So
performing a long async operation (generating the ssh key + adding it as
a deploy key) from initState, results in it being called twice. I'm not
sure where the state is initialized twice.

Even adding a unique key to the widget did not help. In the end, I've
just moved the code to its parent widget.