Pro Mode [Android]: Verify the purchase locally and without RevenueCat
We're using RevenueCat to manage the subscriptions. In my case in
Android, it gives me data saying I'm missing a subscription even though
I clearly have it (though in Trial mode).

The subscription token isn't really being verified right now. Ideally
this needs to be done on the server side. This is a good first start to
get rid of RevenueCat.

Related to #171