Add a YearlyPurchaseWidget - disabled for now
Abstract out the PurchaseWidget and use it for both monthly and yearly
purchases. The yearly purchases hasn't been enabled so far, as the IAPs
still need to be approved on iOS and the PageView widget I was using to
switch between the two options isn't giving me what I want.