ios: Use git_fetch and git_merge instead of pull
Now with the latest git_bindings we no longer use git_pull which would
do a fetch + merge. It's good that these two actions are separated as
I can then easily replace the git_merge with a smarter version
implemented in git.

Also, that way in the future, the git_bindings plugin will only be used
for git_fetch and git_push for SSH. Everything else can be handled by
dart_git.