Editors: Remove the padding
This is weird. In some themes there in an internal padding, and in some
themes, there in none. I didn't think themes would affect this.

Also isDense and isCollapsed work in strange ways. I wish this was
properly documented.

I basically want to remove the left padding, and insert a normal top and
bottom padding. Is that too much to ask?