Remove firebase analytics
Sadly, even when firebase analytics is disabled, it still sends some
data to Google's servers and there seems to be no way to disable it.
GitJournal gives its users an option to disable all analytics, which
clearly can't be implemented as long as we use Firebase.

Therefore there are 2 options -
1. Remove the option of disabling all analytics
2. Remove firebase analytics.

I'm choosing option (2), as it seems wrong to remove the option to
disable analytics, as I know a number of users who have disabled
analytics.

Currently I'm mostly using analytics to measure how well the marketting
is doing, and to see which settings are being used. Disabling the
analytics would make this much much harder, but I guess it's a greater
push for me to find an alternative / create one.

Fixes #506
Related to #105