StripMarkdownFormatting: Only consider '* ' as lists
If we let the space be optional then this picks up bold/italics as well.
Ideally, the stripMarkdown formatting could be made smarter, but I would
prefer to just rid of it in the future, and use the proper markdown
parser.

The only thing that is stopping me right now is performance. So in the
future when the stripped version is cached, this should be fine.

Fixes #420