Fix assertion error in metadata screen (#619)
Reason for changes

In debug/dev mode it is impossible to enter metadata edition
screen since it results in an assertion error. The production
flavour doesn't have this problem, because asserts are removed.
The error is caused by empty repo and file paths that in the Note
constructor inserted by metadata screen build method. This screen
needs a note just to show effects of a user configuration.

What changed?

I added a new factory method to File in order to avoid assertion
checks.