Stop versioning the Podfile lock file
It's a pain as one cannot just upgrade a version in the pubspec but it
also needs to then be updated on that specific platform. Considering we
already have a lockfile from Flutter, I don't see what the lockfile from
Pod adds.

We still need a Podfile as that's where we can configure the build and
add permissions.