Always use dart-git for merges
I'm going to be moving away from libgit2 for the next release, and while
my own dart-git's merging is quite bad in comparison to libgit2, it's
the only way to move forward without libgit2.

The cross-compilation and all is too much of a pain for me to continue
using it. I haven't had an iOS release in over a year because of it.
Plus it sucks all the fun out this project.

Fixes #934